requirements.txt
aniso8601==7.0.0
appnope==0.1.0
astroid==2.2.3
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.6.0
blessings==1.7
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
clarifai==2.6.1
Click==7.0
click-default-group==1.2
colorama==0.4.1
config42==0.3.2
configparser==3.7.3
cycler==0.10.0
decorator==4.4.0
dlib==19.17.99
EasyProcess==0.2.5
emoji==0.5.1
face-recognition==1.2.3
face-recognition-models==0.3.0
fake-useragent==0.1.11
Flask==1.0.3
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
future==0.17.1
googleapis-common-protos==1.5.8
grpcio==1.19.0
html5lib==1.0.1
idna==2.6
instabot-py==0.4.10
instaloader==4.2.8
instapy==0.4.1
instapy-chromedriver==2.46
ipython==7.4.0
ipython-genutils==0.2.0
isort==4.3.10
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.4
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
MeaningCloud-python==1.1.1
MechanicalSoup==0.11.0
numpy==1.16.2
pandas==0.24.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
plyer==1.4.0
profiling==0.1.3
prompt-toolkit==2.0.9
protobuf==3.7.0
psutil==5.6.1
ptyprocess==0.6.0
pycodestyle==2.5.0
pydotplus==2.0.2
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
PySimpleGUI==4.0.0
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2018.9
PyVirtualDisplay==0.2.1
PyYAML==5.1
regex==2019.3.9
requests==2.22.0
requests-futures==1.0.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
selenium==3.141.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.2
spotipy==2.4.4
SQLAlchemy==1.3.4
stem==1.7.1
TBB==0.1
torrequest==0.1.0
traitlets==4.3.2
typed-ast==1.3.1
urllib3==1.22
urwid==2.0.1
valuedispatch==0.0.1
vprof==0.37.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
wrapt==1.11.1
